Checking In
So, I had issues with this.  I had to build the site in a location that could run my .php since I used PHP to get the nav bar to be an include file.  But I didn't know how to move the whole project with Git. Anyway, so that's why there is just one large last check in.